Refactor configuration for simulation script.
